// Modulo_4/Java/mi-app-ute/src/navigation/types.ts
export type RootStackParamList = {
  Home: undefined;
  Details: { id: string; title: string;descripcion:string ;marca:string};
  Settings: undefined;
  Focus: undefined;
  LoginForm: undefined;
  Pantalla: undefined;
  Products: undefined;
  Profile: { userId: string; name: string; apellido:string };
  DisplayName :undefined;
  Fuerza:undefined;
};